+++ src/sdk/queries.ts
@@ -4,6 +4,6 @@ export const QUERY_ORG_REPOS_AGGREGATE_VARIABLES: Record<string, unknown> = {
   pageSize: 100
 }
 
-export const QUERY_ORG_REPOS_AGGREGATE = `
-  
-  `
+// export const QUERY_ORG_REPOS_AGGREGATE = `
+
+//   `